Restructure @carto/core + @carto/ui
