+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 dist
+custom-elements.json
 
 # https://yarnpkg.com/getting-started/qa#which-files-should-be-gitignored
 .pnp.*
