+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing
+
+## Quickstart
+
+TODO
+
+## Dependencies
+
+New `dependencies` are added to individual packages in `packages/*` as required. New `devDependencies` are added to the workspace root `package.json` only. The `devDependencies` for all `packages/*/package.json` should be empty.
