+++ examples/pure-js/pure-js.ts
@@ -6,8 +6,8 @@ import {
   Filter,
   vectorTableSource,
   VectorTableSourceResponse,
-  Widget,
-} from '../../';
+} from '@carto/core';
+import type {Widget} from '@carto/ui';
 
 /**************************************************************************
  * REACTIVE STATE
