+++ package.json
@@ -1,41 +1,16 @@
 {
-  "name": "@carto/widgets",
-  "version": "0.0.0",
-  "description": "TODO",
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "author": "Don McCurdy <donmccurdy@carto.com>",
   "license": "UNLICENSED",
-  "keywords": [
-    "web-components",
-    "lit-element",
-    "typescript",
-    "lit"
-  ],
   "type": "module",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./react": {
-      "types": "./dist/react.d.ts",
-      "default": "./dist/react.js"
-    }
-  },
   "scripts": {
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "dev": "concurrently \"yarn build:watch\" \"vite --config examples/vite.config.js --open\"",
-    "clean": "rimraf dist/*",
-    "lint": "npm run lint:lit-analyzer",
-    "lint:lit-analyzer": "lit-analyzer",
+    "build": "lerna run build",
+    "dev": "concurrently \"yarn build -- --watch\" \"vite --config examples/vite.config.js --open\"",
+    "clean": "rimraf packages/*/dist/*",
     "format": "prettier \"**/*.{cjs,html,js,json,md,ts}\" --ignore-path ./.eslintignore --write"
-  },
-  "dependencies": {
-    "@lit/react": "^1.0.5",
-    "@lit/task": "^1.0.0",
-    "echarts": "^5.5.0",
-    "lit": "^3.0.0"
   },
   "devDependencies": {
     "@deck.gl/aggregation-layers": "^9.0.14",
@@ -57,6 +32,7 @@
     "@vitejs/plugin-vue": "^5.0.5",
     "@webcomponents/webcomponentsjs": "^2.8.0",
     "concurrently": "^8.2.2",
+    "lerna": "^8.1.3",
     "lit-analyzer": "^1.2.1",
     "maplibre-gl": "^4.1.3",
     "prettier": "^2.6.2",
@@ -69,6 +45,5 @@
     "vite": "^5.2.10",
     "vue": "^3.4.27"
   },
-  "customElements": "custom-elements.json",
   "packageManager": "yarn@4.2.2"
 }
