+++ packages/core/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@carto/core",
+  "packageManager": "yarn@4.2.2",
+  "author": "Don McCurdy <donmccurdy@carto.com>",
+  "license": "UNLICENSED",
+  "type": "module",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "lint": "echo 'lint not implemented' && exit 1"
+  }
+}
