+++ packages/core/src/constants.ts
@@ -1,0 +1,3 @@
+import {FilterTypes} from './vendor/deck-carto.js';
+
+export const FILTER_TYPES = new Set(Object.values(FilterTypes));
