+++ packages/core/src/index.ts
@@ -1,8 +1,7 @@
-export * from './widgets/index.js';
 export * from './sources/index.js';
 
 // TODO: Clean up and consolidate shared constants and types.
-export {FilterEvent} from './types.js';
+export * from './types.js';
 export type {Filter} from './vendor/deck-carto.js';
 export {
   MAP_TYPES,
