+++ packages/core/src/sources/base-widget-source.ts
