+++ packages/core/src/sources/index.ts
@@ -1,4 +1,4 @@
-export {WidgetSource} from './base-widget-source';
+export {WidgetSource} from './base-widget-source.js';
 export * from './query-widget-source.js';
 export * from './table-widget-source.js';
 export * from './wrappers.js';
