+++ packages/core/src/types.ts
@@ -1,7 +1,7 @@
 import {AggregationTypes} from './vendor/carto-constants.js';
-import {$TODO} from './types-internal';
-import {SpatialFilter} from './vendor/carto-react-api';
-import {Filter} from './vendor/deck-carto';
+import {$TODO} from './types-internal.js';
+import {SpatialFilter} from './vendor/carto-react-api.js';
+import {Filter} from './vendor/deck-carto.js';
 
 /******************************************************************************
  * FILTERS
