+++ packages/core/src/utils.ts
@@ -1,11 +1,6 @@
-import {MapViewState, WebMercatorViewport} from '@deck.gl/core';
-import {Filter, FilterTypes} from './vendor/deck-carto.js';
 import {FILTER_TYPES} from './constants.js';
 import {$TODO} from './types-internal.js';
-
-export function sleep(ms: number) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
+import {Filter, FilterTypes} from './vendor/deck-carto.js';
 
 export function getWidgetFilters(
   owner?: string,
@@ -28,20 +23,4 @@ export function getWidgetFilters(
   }
 
   return widgetFilters;
-}
-
-export function getSpatialFilter(viewState: MapViewState): any {
-  const viewport = new WebMercatorViewport(viewState);
-  return {
-    type: 'Polygon',
-    coordinates: [
-      [
-        viewport.unproject([0, 0]),
-        viewport.unproject([viewport.width, 0]),
-        viewport.unproject([viewport.width, viewport.height]),
-        viewport.unproject([0, viewport.height]),
-        viewport.unproject([0, 0]),
-      ],
-    ],
-  };
 }
