+++ packages/core/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.json",
+	"include": ["src/**/*", "../global.d.ts"],
+	"compilerOptions": {
+		"outDir": "dist",
+		"allowJs": true // TODO
+	}
+}
