+++ packages/ui/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@carto/ui",
+  "packageManager": "yarn@4.2.2",
+  "author": "Don McCurdy <donmccurdy@carto.com>",
+  "license": "UNLICENSED",
+  "type": "module",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./react": {
+      "types": "./dist/react.d.ts",
+      "default": "./dist/react.js"
+    }
+  },
+  "customElements": "custom-elements.json",
+  "scripts": {
+    "build": "tsc",
+    "lint": "lit-analyzer"
+  },
+  "dependencies": {
+    "@carto/core": "workspace:^",
+    "@lit/react": "^1.0.5",
+    "@lit/task": "^1.0.0",
+    "echarts": "^5.5.0",
+    "lit": "^3.0.0"
+  },
+  "optionalDependencies": {
+    "@types/react": ">=18.0.0",
+    "react": ">=18.0.0"
+  },
+  "peerDependencies": {
+    "@deck.gl/core": ">=9.0.0",
+    "@types/react": ">=18.0.0",
+    "react": ">=18.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@types/react": {
+      "optional": true
+    },
+    "react": {
+      "optional": true
+    }
+  }
+}
