+++ packages/ui/src/constants.ts
@@ -1,0 +1,1 @@
+export const DEBOUNCE_TIME_MS = 200;
