+++ packages/ui/src/widgets/scatter-widget.ts
@@ -2,15 +2,15 @@ import {LitElement, html} from 'lit';
 import {Task} from '@lit/task';
 import {customElement, property, state} from 'lit/decorators.js';
 import {Ref, createRef, ref} from 'lit/directives/ref.js';
-import {DEBOUNCE_TIME_MS} from '../constants';
-import {getSpatialFilter, sleep} from '../utils';
+import {cache} from 'lit/directives/cache.js';
 import * as echarts from 'echarts';
 import {TaskStatus} from '@lit/task';
-import {DEFAULT_TEXT_STYLE, WIDGET_BASE_CSS} from './styles';
-import {cache} from 'lit/directives/cache.js';
-import {AggregationTypes} from '../vendor/carto-constants';
-import {WidgetSource} from '../sources/index.js';
+import {AggregationTypes, WidgetSource} from '@carto/core';
 import {MapViewState} from '@deck.gl/core';
+
+import {DEBOUNCE_TIME_MS} from '../constants.js';
+import {getSpatialFilter, sleep} from '../utils.js';
+import {DEFAULT_TEXT_STYLE, WIDGET_BASE_CSS} from './styles.js';
 
 const DEFAULT_SCATTER_GRID = {
   left: '90px',
