--- src/constants.ts
@@ -1,4 +1,0 @@
-import {FilterTypes} from './vendor/deck-carto';
-
-export const DEBOUNCE_TIME_MS = 200;
-export const FILTER_TYPES = new Set(Object.values(FilterTypes));
