+++ tsconfig.json
@@ -1,31 +1,27 @@
 {
   "include": [
-    "src/**/*.ts",
-    "examples/*.{svelte,ts,tsx,vue}"
+    "packages/*/src/**/*.ts",
+    "examples/**/*.{svelte,ts,tsx,vue}"
   ],
   "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "target": "es2021",
-    "module": "es2020",
+    "baseUrl": "./packages",
+    "paths": {
+      "@carto/core": ["./packages/core/"],
+      "@carto/ui": ["./packages/ui/"]
+    },
+    "target": "es2022",
+    "module": "es2022",
     "moduleResolution": "bundler",
-    "lib": ["es2021", "DOM", "DOM.Iterable"],
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "lib": ["es2022", "DOM", "DOM.Iterable"],
     "types": ["vite/client"],
     "jsx": "react",
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
     "inlineSources": true,
-    "allowJs": true, // TODO
-    "skipLibCheck": true,
     "strict": true,
-    "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
-    "plugins": [
-      {
-        "name": "ts-lit-plugin",
-        "strict": true
-      }
-    ]
+    "skipLibCheck": true
   }
 }
